canopeum_backend/requirements.txt
alabaster==0.7.16
asgiref==3.7.2
attrs==23.2.0
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djlint==1.34.1
docutils==0.20.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.6
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.3
jsbeautifier==1.15.1
json5==0.9.22
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.0.1
packaging==23.2
parsimonious==0.10.0
pathspec==0.12.1
Pygments==2.17.2
pyjson5==1.6.6
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
setuptools==69.2.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.2.5
sphinx-js==3.2.2
sphinx-multiproject==1.0.0rc1
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.4.4
tqdm==4.66.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
